Linking attributes
Colour data now gets passed into the VAO so that more than one colour can be used in the shaders.

An in and out vec3 called colour was created in the vertex shader. This is then read in by the fragment shader and used to draw a gradient on the triangle primitives