GameObjects and Pool Table and Ball
Models were added for a Pool Table and Pool Ball.

A GameObject class was made to be a wrapper around the Model class, feeding it position and other values. This extra layer of abstraction means that the gameplay and graphics can be change independently and still work with each other